Add `Counted` decorator for `Operations`. Implements #23
The `Counted` decorator sets expected number of affected rows. If the actual number of affected rows doesn't match the given number the entire transaction will fail. This can be used with Update, Delete and Assert Operations.
This also adds an `Empty` decorator which is a special case of `Counted` with an expected number of affected rows of zero.