Add AggregationExceptions. Implements #10
To use it just create a `Link`, `Split` or `AutoAggregate` operation for two specific contacts.

In addition this adds bulk operations to link or split a raw contact to/from multiple other raw contacts.

This commit also adds the TransactionContext to `RowReference.builderWithReferenceData(…)` in order to be able to add references to virtual RowSnapshots.