Add ReferringTo predicate. Implements #46
This commit adds TransactionContext to Predicates in order to resolve related rows dynamically at operation build time.
BaseView currently passes an EmptyTransactionContext which keeps the current behavior. In future versions it might be useful
to add a TransactionContext to `View.rows(…)` to be able to query by rows which have been inserted in a previous transaction.