Fix reference resolution. Fixes #64
Only add a rowReference to the temporary and result transaction contexts if it's virtual and can not be resolved otherwise.
Before this, an update or assert of an insert operation resulted in an updated transaction context with follow up resolve operations returning a reference to the update or assert operation, which can not work.