Introduce dedicated Projection type. Implements #134 (#137)
This commit adds a dedicated `Projection` type which replaces plain String arrays. This type allows to declare projections for views in a convenient manner.

This also changes the projection from being a `View` property to being a `RowSet` property which is passed to the `View` when it's actually queried. This way a single `View` per table is sufficient.